Some cleaning up   - Removed unnessersary include dirs from Gene & Gene.Sandbox   - Removed some warnings from Gene core   - Placed all freetype-gl code into a namespace

+++ Gene/Engine/Private/Math/Quaternion.cc
@@ -11,7 +11,7 @@ Quaternion::Quaternion(const Vector3& n, float a)
 
 	W = ::cosf(halfA);
 
-	X = n.X * ::sin(halfA);
-	Y = n.Y * ::sin(halfA);
-	Z = n.Z * ::sin(halfA);
+	X = n.X * ::sinf(halfA);
+	Y = n.Y * ::sinf(halfA);
+	Z = n.Z * ::sinf(halfA);
 }
